+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/window/Windows.java
@@ -65,8 +65,8 @@ public final class Windows {
 	 * @param screenSizeRatio the screen size ratio
 	 * @throws IllegalArgumentException in case ratio is not between 0 and 1
 	 */
-	public static void resizeWindow(Window window, double screenSizeRatio) {
-		resizeWindow(window, screenSizeRatio, null, null);
+	public static void resize(Window window, double screenSizeRatio) {
+		resize(window, screenSizeRatio, null, null);
 	}
 
 	/**
@@ -78,7 +78,7 @@ public final class Windows {
 	 * @param maximumSize the maximum size, may be null
 	 * @throws IllegalArgumentException in case ratio is not between 0 and 1
 	 */
-	public static void resizeWindow(Window window, double screenSizeRatio, Dimension minimumSize, Dimension maximumSize) {
+	public static void resize(Window window, double screenSizeRatio, Dimension minimumSize, Dimension maximumSize) {
 		Dimension ratioSize = screenSizeRatio(screenSizeRatio);
 		if (minimumSize != null) {
 			ratioSize.setSize(Math.max(minimumSize.width, ratioSize.width), Math.max(minimumSize.height, ratioSize.height));